update documentation on VCU code and remove beaglebone code
