+++ vcu/Cube-F7-Src-respin/2025_VCU.ioc
@@ -65,7 +65,7 @@ Dma.UART4_RX.1.RequestParameters=Instance,Direction,PeriphInc,MemInc,PeriphDataA
 FREERTOS.FootprintOK=true
 FREERTOS.INCLUDE_vTaskDelayUntil=1
 FREERTOS.IPParameters=Tasks01,FootprintOK,configTOTAL_HEAP_SIZE,configUSE_TIMERS,configGENERATE_RUN_TIME_STATS,configUSE_TRACE_FACILITY,configTIMER_TASK_PRIORITY,configTIMER_QUEUE_LENGTH,configTIMER_TASK_STACK_DEPTH,configCHECK_FOR_STACK_OVERFLOW,configUSE_STATS_FORMATTING_FUNCTIONS,configUSE_COUNTING_SEMAPHORES,INCLUDE_vTaskDelayUntil
-FREERTOS.Tasks01=driveByWire,3,1000,driveByWireTask,As weak,NULL,Dynamic,NULL,NULL;mainTask,0,1000,mainTaskFunction,As external,NULL,Dynamic,NULL,NULL;printTaskName,-2,1000,printTask,As external,NULL,Dynamic,NULL,NULL;cliTaskName,-2,1000,cliTask,As external,NULL,Dynamic,NULL,NULL;watchdogTaskNam,3,1000,watchdogTask,As external,NULL,Dynamic,NULL,NULL;canSendTask,3,1000,canTask,As external,NULL,Dynamic,NULL,NULL;canPublish,0,1000,canPublishTask,As external,NULL,Dynamic,NULL,NULL;beaglebone,0,1000,bbTask,As external,NULL,Dynamic,NULL,NULL;enduranceMode,0,10000,enduranceModeTask,As external,NULL,Dynamic,NULL,NULL;tractionControl,0,1024,tractionControlTask,As external,NULL,Dynamic,NULL,NULL;throttlePolling,3,1000,throttlePollingTask,As external,NULL,Dynamic,NULL,NULL
+FREERTOS.Tasks01=driveByWire,3,1000,driveByWireTask,As weak,NULL,Dynamic,NULL,NULL;mainTask,0,1000,mainTaskFunction,As external,NULL,Dynamic,NULL,NULL;printTaskName,-2,1000,printTask,As external,NULL,Dynamic,NULL,NULL;cliTaskName,-2,1000,cliTask,As external,NULL,Dynamic,NULL,NULL;watchdogTaskNam,3,1000,watchdogTask,As external,NULL,Dynamic,NULL,NULL;canSendTask,3,1000,canTask,As external,NULL,Dynamic,NULL,NULL;canPublish,0,1000,canPublishTask,As external,NULL,Dynamic,NULL,NULL;enduranceMode,0,10000,enduranceModeTask,As external,NULL,Dynamic,NULL,NULL;tractionControl,0,1024,tractionControlTask,As external,NULL,Dynamic,NULL,NULL;throttlePolling,3,1000,throttlePollingTask,As external,NULL,Dynamic,NULL,NULL
 FREERTOS.configCHECK_FOR_STACK_OVERFLOW=2
 FREERTOS.configGENERATE_RUN_TIME_STATS=1
 FREERTOS.configTIMER_QUEUE_LENGTH=4
