+++ vcu/Cube-F7-Src-respin/Inc/main.h
@@ -32,6 +32,7 @@ extern "C" {
 
 /* Private includes ----------------------------------------------------------*/
 /* USER CODE BEGIN Includes */
+#include "generalErrorHandler.h"
 /* USER CODE END Includes */
 
 /* Exported types ------------------------------------------------------------*/
