--- vcu/Inc/beaglebone.h
@@ -1,10 +1,0 @@
-#ifndef BEAGLEBONE_H
-
-#define BEAGLEBONE_H
-
-#include "bsp.h"
-
-HAL_StatusTypeDef beaglebonePower(bool enable);
-HAL_StatusTypeDef beagleboneOff();
-
-#endif /* end of include guard: BEAGLEBONE_H */
