+++ vcu/Src/mathUtils.c
@@ -9,6 +9,18 @@ float min(float a, float b)
     }
 }
 
+int map_range(int in, int low, int high, int low_out, int high_out) {
+    if (in < low) {
+        in = low;
+    } else if (in > high) {
+        in = high;
+    }
+    int in_range = high - low;
+    int out_range = high_out - low_out;
+
+    return (in - low) * out_range / in_range + low_out;
+}
+
 float map_range_float(float in, float low, float high, float low_out, float high_out)
 {
     if (in < low) {
